requirements.txt
fastapi[standard]
uvicorn
passlib[bcrypt]
sqlalchemy
python-jose[cryptography]